requirements.txt

asgiref==3.6.0
build==0.10.0
charset-normalizer==3.1.0
colorama==0.4.6
crashtest==0.4.1
distlib==0.3.6
dj-database-url==1.2.0
Django==3.2.8
django-cors-headers==3.14.0
django-heroku==0.3.1
djangorestframework==3.14.0
gunicorn==20.1.0
installer==0.7.0
numpy==1.24.2
packaging==23.0
pexpect==4.8.0
pkginfo==1.9.6
platformdirs==2.6.2
PyMySQL==1.0.2
pyproject_hooks==1.0.0
pyrsistent==0.19.3
python-dotenv==1.0.0
typing_extensions==4.5.0
virtualenv==20.21.0
webencodings==0.5.1
whitenoise==6.4.0
zipp==3.15.0
